// src/app/interfaces/user-info.ts
export interface UserInfo {
    id: number;
    login: string;
    name: string;
    email: string;
    avatar_url: string;
    blog: string;
    twitter_username: string;
    company: string;
    location: string;
    hireable: boolean;
    bio: string;
    followers: number;
    following: number;
    created_at: string;
}